Implement deleting messages for user
- only deletes user's view of message, other users can still see it if they haven't deleted it for themselves

+++ peer_support/models/message_model.py
@@ -7,10 +7,16 @@ class Message(models.Model):
     sender = models.ForeignKey(User,null=True,on_delete=models.SET_NULL,unique=False)
     content = models.CharField(max_length=100)
     send_time = models.DateTimeField(default=timezone.now)
+    visible_to = models.ManyToManyField(User, blank=True,related_name='visible_to')
+    read_by = models.ManyToManyField(User, blank=True,related_name = 'read_by')
 
-    def delete(self):
+    def delete(self,users):
         """Delete message"""
-        Message.objects.filter(pk=self.pk).delete()
+        self.visible_to.remove(users)
+        print(self.visible_to.all())
+        if self.visible_to.count() == 0:
+            print('wow')
+            Message.objects.filter(pk=self.pk).delete()
 
     def __str__(self):
         """Return a string representing the message"""
